+++ src/main/java/com/project/service/orderService.java
@@ -2,6 +2,7 @@ package com.project.service;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.Random;
 import java.util.stream.Collectors;
 
 import org.modelmapper.ModelMapper;
@@ -10,7 +11,7 @@ import org.springframework.stereotype.Service;
 
 import com.project.DTO.orderDTO;
 import com.project.entities.orderTable;
-import com.project.exceptions.AccountNotFoundExceptions;
+import com.project.exceptions.IDNotFoundExceptions;
 import com.project.repo.OrdersRepo;
 
 
@@ -42,14 +43,37 @@ public orderDTO create(orderTable entity) {
       return this.MapToDTO(entity);
  }
 
+public String createUnique(orderTable entity) {
+	Random rand = new Random();
+	long range = 9999999L;
+	long number = (long)(rand.nextDouble()*range);
+    this.repo.saveOrdersBySQL(number, entity.getOrderQuantity(), entity.getProcess());
+    return "Order Created"; 
+}
+
 
 public boolean delete(long id) {
-	  orderTable ent = this.repo.findById(id).orElseThrow(AccountNotFoundExceptions::new);
+	if(!this.repo.existsById(id)) {
+		throw new IDNotFoundExceptions();
+	}
+	else {
+
       this.repo.deleteById(id);
       boolean isExist = this.repo.existsById(id);
       return isExist; 
-     
+	}
  }
+public boolean deleteUniqueID(long id) {
+	if(this.repo.findAllUniqueBySQL(id) == null) {
+		throw new IDNotFoundExceptions();
+	}
+	else {
+    this.repo.deleteOrdersBySQLUniqueID(id);
+    boolean isExist = true;
+    return isExist; 
+   }
+}
+
 
 
 public orderDTO update(long id, orderTable entity) {
@@ -57,9 +81,16 @@ public orderDTO update(long id, orderTable entity) {
 	Optional<orderTable> existingOptional = this.repo.findById(id);
 	orderTable ent = existingOptional.get();
 	ent.setOrderQuantity(entity.getOrderQuantity());
-	ent.isProcessing();
+	ent.setProcess(entity.getProcess());
 	entity = this.repo.save(ent);
 	return this.MapToDTO(entity);
  }
+/*
+public void updateUnique(long id, orderTable entity) {
+	
+	this.repo.findAllUniqueBySQL(id);
+	this.repo.updateOrdersBySQL(entity.getOrderQuantity() , entity.getProcess(), entity.getOrderUniqueID());
+	//return this.MapToDTO(entity);
+ }*/
 
 }
